backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.17.1",
    "asyncpg>=0.30.0",
    "cryptography>=46.0.3",
    "fastapi>=0.121.0",
    "google-generativeai>=0.8.5",
    "passlib>=1.7.4",
    "pydantic-settings>=2.11.0",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "qdrant-client>=1.15.1",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
